Set Character Code Size To One Byte
Set character code size to one byte.

+++ HuffmanCompressor/src/CharacterCountResult.java
@@ -1,15 +1,15 @@
 import java.util.Map;
 
 public class CharacterCountResult {
-    private final Map<Integer, Integer> characterCount;
+    private final Map<Byte, Integer> characterCount;
     private final int totalCharacterCount;
 
-    CharacterCountResult(Map<Integer, Integer> characterCount, int totalCharacterCount) {
+    CharacterCountResult(Map<Byte, Integer> characterCount, int totalCharacterCount) {
         this.characterCount = characterCount;
         this.totalCharacterCount = totalCharacterCount;
     }
 
-    public Map<Integer, Integer> getCharacterCount() {
+    public Map<Byte, Integer> getCharacterCount() {
         return characterCount;
     }
